refactor(skills): de-package scienceskillscommon into plain script modules
- move http_client.py + jats.py into skills/literature-review/scripts/ as
  sibling modules; scripts import them directly (uv run puts the script dir
  on sys.path). Tests load by path and sys.path-insert the scripts dir.
- delete the scienceskillscommon package: dir, pyproject, __init__, and the
  stub SKILL.md that made it a phantom 'co-researcher:scienceskillscommon'
  skill. Drop scienceskillscommon from every PEP723 header + uv.sources.
- removes the foreign 'science-skills' name, the wheel build/remap, and the
  stale-wheel --reinstall gotcha (edits are now always live)
- prune 3 obsolete ProjectContext pitfalls; 45 tests green, live verified
