--- skills/scienceskillscommon/__init__.py
@@ -1,10 +1,0 @@
-# MIT License
-#
-# Copyright (c) 2026 Poe Poe / co-researcher contributors
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction. THE SOFTWARE IS PROVIDED "AS IS", WITHOUT
-# WARRANTY OF ANY KIND. See the MIT License for details.
-
-"""Package marker for science_skills.scienceskillscommon."""
