--- skills/scienceskillscommon/pyproject.toml
@@ -1,14 +1,0 @@
-[project]
-name = "scienceskillscommon"
-version = "0.2.0"
-license = "MIT"
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build.targets.wheel]
-packages = ["."]
-
-[tool.hatch.build.targets.wheel.sources]
-"." = "science_skills/scienceskillscommon"
