+++ tests/test_openalex_cli.py
@@ -2,11 +2,8 @@
 # requires-python = ">=3.10"
 # dependencies = [
 #   "pytest",
-#   "scienceskillscommon",
 #   "python-dotenv",
 # ]
-# [tool.uv.sources]
-# scienceskillscommon = { path = "../skills/scienceskillscommon" }
 # ///
 
 import argparse
@@ -17,10 +14,10 @@ import sys
 
 import pytest
 
-_SCRIPT = (
-    pathlib.Path(__file__).resolve().parent.parent
-    / "skills/literature-review/scripts/openalex_cli.py"
-)
+_SCRIPTS = (pathlib.Path(__file__).resolve().parent.parent
+            / "skills/literature-review/scripts")
+sys.path.insert(0, str(_SCRIPTS))
+_SCRIPT = _SCRIPTS / "openalex_cli.py"
 _spec = importlib.util.spec_from_file_location("openalex_cli", _SCRIPT)
 openalex_cli = importlib.util.module_from_spec(_spec)
 _spec.loader.exec_module(openalex_cli)
