+++ tests/test_read_paper.py
@@ -2,11 +2,8 @@
 # requires-python = ">=3.10"
 # dependencies = [
 #   "pytest",
-#   "scienceskillscommon",
 #   "pymupdf4llm",
 # ]
-# [tool.uv.sources]
-# scienceskillscommon = { path = "../skills/scienceskillscommon" }
 # ///
 
 import importlib.util
@@ -18,10 +15,10 @@ import urllib.parse
 
 import pytest
 
-_SCRIPT = (
-    pathlib.Path(__file__).resolve().parent.parent
-    / "skills/literature-review/scripts/read_paper.py"
-)
+_SCRIPTS = (pathlib.Path(__file__).resolve().parent.parent
+            / "skills/literature-review/scripts")
+sys.path.insert(0, str(_SCRIPTS))
+_SCRIPT = _SCRIPTS / "read_paper.py"
 _spec = importlib.util.spec_from_file_location("read_paper", _SCRIPT)
 read_paper = importlib.util.module_from_spec(_spec)
 _spec.loader.exec_module(read_paper)
